project_25_email_sender/requirements.txt
certifi==2026.1.4
python-dotenv==1.2.1
